Add HomePage e2e test
HomePage Component:
Verify that the "CreateKeys" and "Login" components are rendered correctly on the page.
Check that elements related to the title, input field, "Submit" button, "Don't have an account?" link, and "Click here" button exist on the page.

Login Component:
Display an error message if the length of the secret key is less than 56 characters.
Display an error message if the secret key does not start with 'S'.
Submit the form with a valid secret key and verify that the error message is not displayed.

Generate Keys Component:
Generate a secret key by clicking the "Click here" button.
Verify that the input field is not empty after generating the secret key.