Add Sign Out test
These test cases serve to verify the logout functionality.
The 1st test case ensures that once logged in, if there is 3 minutes of inactivity, the session will be terminated and the home page will be displayed.
The 2nd test case ensures that when we are inside our wallet and click the 'Sign Out' button, the session will be terminated.