Add accountSender input
Added the form to write your secretKey along with the security conditions for this one. They are similar to validateAccountReceiver only that in this case validateaccountSender evaluates that the written account is equal to your publicKey.
It was also added that if there is an error the 'Send' button is disabled.